casper_push_postinstall_check.sh
This script is to be used in a policy to deploy either a flat DMG
package or a DMG that contains a pkg or mpkg. This verifies the DMG is
valid before installing.
The packaged should be set to Cache in the policy.
The script should be set to execute "After" the package has been cached.